+++ app/themen.py
@@ -2,7 +2,7 @@
 
 import cherrypy
 from mako.template import Template
-from app import database
+from app import database, application
 
 
 class Themen_cl(object):
@@ -15,6 +15,7 @@ class Themen_cl(object):
 
 	@cherrypy.expose
 	def index(self):
+		cherrypy.Application.user.user_logged_in()
 		themen = self.db.load_themen()
 
 		template = Template(filename="content/themen/index.html")
@@ -22,12 +23,20 @@ class Themen_cl(object):
 
 	@cherrypy.expose
 	def delete(self, themen_id):
+		cherrypy.Application.user.user_logged_in()
+		if not cherrypy.Application.user.is_admin():
+			return cherrypy.Application.view.error("403")
+
 		self.db.delete_thema(themen_id)
 
 		raise cherrypy.HTTPRedirect("/themen/index")
 
 	@cherrypy.expose
 	def create(self, **kwargs):
+		cherrypy.Application.user.user_logged_in()
+		if not cherrypy.Application.user.is_admin():
+			return cherrypy.Application.view.error("403")
+
 		if "name" in kwargs and "beschreibung" in kwargs:
 			self.db.create_thema(kwargs["name"], kwargs["beschreibung"])
 			raise cherrypy.HTTPRedirect("/themen/index")
